refactor(labyrinth): extract the space-time reachability subsystem into aether-labyrinth
Move the reachability solver, the corridor-graph + resolution-depth
analysis, the counterfactual / traffic / escapability passes, the nine
certifier #[transform]s, the aether.trajectory recorder cap, and the
#1860 baseline harness out of aether-capabilities into a new
aether-labyrinth crate. The crate reads as "the reachability primitive
plus what's built on it": a lib.rs module map plus one-line rustdoc per
transform and the recorder cap.

The transforms are pure (aether-data + aether-kinds only), so the default
crate is transforms-only and aether-mcp deps it for the link-time
#[transform] inventory describe_transforms reads — without pulling the
substrate / wgpu / wasmtime stack (verified absent from mcp's normal dep
graph). TrajectoryRecorderCapability implements NativeActor, so it rides
behind the native feature; aether-substrate-bundle enables it and the
dependency edge keeps the certifier transforms registering in the
headless TransformRegistry. The mat4_apply generic transform stays in
aether-capabilities; the recorder cap's const NAMESPACE = "aether.trajectory"
is unchanged (the mailbox id is hash-derived from it, ADR-0099).

Beyond the plan's named cores, counterfactual.rs / traffic.rs /
escapability.rs move too: they are the pure cores the explicitly-listed
solve_counterfactual / aggregate_traffic / realization transforms hard
depend on, used nowhere else, so the subsystem moves whole.
