+++ crates/aether-labyrinth/src/escapability.rs
