+++ crates/aether-labyrinth/src/trajectory.rs
