+++ crates/aether-substrate-bundle/Cargo.toml
@@ -67,6 +67,12 @@ path = "src/bin/perf-plot.rs"
 [dependencies]
 aether-substrate = { path = "../aether-substrate", features = ["render"] }
 aether-capabilities = { path = "../aether-capabilities", features = ["render-native", "audio-native", "text-native", "ui-native"] }
+# The space-time reachability subsystem (issue 1908). `native` pulls the
+# `TrajectoryRecorderCapability` the chassis registers; the dependency edge
+# is also what links its certifier `#[transform]`s into the headless
+# binary's `TransformRegistry::from_inventory` (a dropped edge silently
+# de-registers them with no compile error).
+aether-labyrinth = { path = "../aether-labyrinth", features = ["native"] }
 aether-actor = { path = "../aether-actor" }
 aether-data = { path = "../aether-data" }
 aether-kinds = { path = "../aether-kinds" }
