+++ crates/aether-substrate-bundle/src/bin/headless.rs
@@ -8,6 +8,17 @@
 // (ADR-0090 §4 / e2).
 #![allow(clippy::print_stdout)]
 
+// Force-link `aether-labyrinth` into this engine binary so its certifier
+// `#[transform]`s register in the link-time `aether_data::transforms()`
+// inventory the headless chassis's `DagCapability` builds its
+// `TransformRegistry` from (issue 1908). The bundle lib references only the
+// recorder cap (a different codegen member), so without this the linker
+// drops the transforms member and a reachability DAG fails validation with
+// no compile error. `as _` is the side-effect-linkage form, and the
+// directive must live in this final-artifact root — it does not propagate
+// in from the bundle lib's own `extern crate`.
+extern crate aether_labyrinth as _;
+
 use aether_substrate::Chassis;
 use aether_substrate_bundle::chassis_config_dump;
 use aether_substrate_bundle::cli::HeadlessCli;
